+++ Doolda/Doolda/Service/Persistences/FileManagerPersistenceService.swift
@@ -8,6 +8,20 @@
 import Combine
 import Foundation
 
+enum FileDocuments {
+    typealias RawValue = URL?
+
+    case temporary
+    case cache
+
+    var rawValue: RawValue {
+        switch self {
+        case .temporary: return FileManager.default.temporaryDirectory
+        case .cache: return FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first
+        }
+    }
+}
+
 enum FileManagerPersistenceServiceError: LocalizedError {
     case failedToSaveFile
     case failedToFetchFile
